app/requirements.txt
boto3
Flask
Flask-SQLAlchemy
PyMySQL
python-dotenv